--- pafi/main.hpp
@@ -1,26 +1,0 @@
-#include <iostream>
-#include <sstream>
-#include <iostream>
-#include <iomanip>
-
-#include <mpi.h>
-#include <cstdio>
-#include <cstdlib>
-#include <cstring>
-#include <vector>
-#include <string>
-#include <array>
-#include <map>
-#include <cmath>
-#include <ctime>
-
-
-
-// PAFI files
-#include "Constants.hpp"
-#include "Parser.hpp"
-#include "Spline.hpp"
-#include "File.hpp"
-
-#include "LAMMPSSimulator.hpp"
-typedef LAMMPSSimulator Simulator;
