+++ pafi/pafi.cpp
